tool/requirements.txt
requests
bs4
reportlab
mysql-connector
nvdlib
colorama